feat: update deploy command to be deploy trigger
